requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pymongo==4.6.0
pyyaml==6.0.1
chromedriver-autoinstaller==0.6.4
pandas==1.5.3
pytest==7.4.3
pytest-cov==4.1.0
